Significant progress with f64 but still unexplained errors with sin/cos.
