+++ src/bin/libmgen/hyperbolic.rs
@@ -1,10 +1,11 @@
 use crate::test::*;
 use proc_macro2::TokenStream;
-use quote::{format_ident, quote};
+use quote::{quote};
+use crate::helpers;
 
 // https://en.wikipedia.org/wiki/Hyperbolic_functions
 pub fn gen_sinh(_num_terms: usize, num_bits: usize) -> TokenStream {
-    let fty = format_ident!("f{}", num_bits);
+    let fty = helpers::get_fty(num_bits);
     quote!(
         fn sinh(x: #fty) -> #fty {
             let a = x.mul_add(std::#fty::consts::LOG2_E, -1.0);
@@ -16,7 +17,7 @@ pub fn gen_sinh(_num_terms: usize, num_bits: usize) -> TokenStream {
 
 // https://en.wikipedia.org/wiki/Hyperbolic_functions
 pub fn gen_cosh(_num_terms: usize, num_bits: usize) -> TokenStream {
-    let fty = format_ident!("f{}", num_bits);
+    let fty = helpers::get_fty(num_bits);
     quote!(
         fn cosh(x: #fty) -> #fty {
             let a = x.mul_add(std::#fty::consts::LOG2_E, -1.0);
@@ -28,7 +29,7 @@ pub fn gen_cosh(_num_terms: usize, num_bits: usize) -> TokenStream {
 
 // https://en.wikipedia.org/wiki/Hyperbolic_functions
 pub fn gen_tanh(_num_terms: usize, num_bits: usize) -> TokenStream {
-    let fty = format_ident!("f{}", num_bits);
+    let fty = helpers::get_fty(num_bits);
     quote!(
         fn tanh(x: #fty) -> #fty {
             let exp2x = exp2(x * (std::#fty::consts::LOG2_E * 2.0));
@@ -39,7 +40,7 @@ pub fn gen_tanh(_num_terms: usize, num_bits: usize) -> TokenStream {
 
 // https://en.wikipedia.org/wiki/Inverse_hyperbolic_functions
 pub fn gen_asinh(_num_terms: usize, num_bits: usize) -> TokenStream {
-    let fty = format_ident!("f{}", num_bits);
+    let fty = helpers::get_fty(num_bits);
     quote!(
         fn asinh(x: #fty) -> #fty {
             ln(x + (x * x + 1.0).sqrt())
@@ -49,7 +50,7 @@ pub fn gen_asinh(_num_terms: usize, num_bits: usize) -> TokenStream {
 
 // https://en.wikipedia.org/wiki/Inverse_hyperbolic_functions
 pub fn gen_acosh(_num_terms: usize, num_bits: usize) -> TokenStream {
-    let fty = format_ident!("f{}", num_bits);
+    let fty = helpers::get_fty(num_bits);
     quote!(
         fn acosh(x: #fty) -> #fty {
             ln(x + (x * x - 1.0).sqrt())
@@ -59,7 +60,7 @@ pub fn gen_acosh(_num_terms: usize, num_bits: usize) -> TokenStream {
 
 // https://en.wikipedia.org/wiki/Inverse_hyperbolic_functions
 pub fn gen_atanh(_num_terms: usize, num_bits: usize) -> TokenStream {
-    let fty = format_ident!("f{}", num_bits);
+    let fty = helpers::get_fty(num_bits);
     quote!(
         fn atanh(x: #fty) -> #fty {
             (ln(1.0 + x) - ln(1.0 - x)) * 0.5
@@ -77,7 +78,7 @@ pub fn gen_hyperbolic(num_bits: usize) -> (TokenStream, TokenStream) {
     let atanh = gen_atanh(7, num_bits);
 
     let bit = (2.0_f64).powi(if num_bits == 32 { 23 } else { 52 });
-    let fty = format_ident!("f{}", num_bits);
+    let fty = helpers::get_fty(num_bits);
 
     let test_cosh = gen_test(
         quote!(test_cosh),
