+++ src/bin/libmgen/log_exp.rs
@@ -1,13 +1,17 @@
 use doctor_syn::Parity;
 use doctor_syn::{expr, name, num_digits_for};
 use proc_macro2::TokenStream;
-use quote::{format_ident, quote};
+use quote::{quote};
+use crate::helpers;
 
 use crate::test::*;
 
 pub fn gen_exp2(num_terms: usize, num_bits: usize) -> TokenStream {
-    let suffix = format!("f{}", num_bits);
-    let fty = format_ident!("f{}", num_bits);
+    let suffix = helpers::get_suffix(num_bits);
+    let fty = helpers::get_fty(num_bits);
+    let uty = helpers::get_uty(num_bits);
+    let one = helpers::get_one(num_bits);
+    let escale = helpers::get_escale(num_bits);
 
     let xmin = -0.5;
     let xmax = 0.5;
@@ -22,7 +26,7 @@ pub fn gen_exp2(num_terms: usize, num_bits: usize) -> TokenStream {
     quote!(
         fn exp2(x: #fty) -> #fty {
             let r = x.round();
-            let mul = #fty::from_bits((r.mul_add(0x00800000 as #fty, 0x3f800000 as #fty)) as u32);
+            let mul = #fty::from_bits((r.mul_add(#escale as #fty, #one as #fty)) as #uty);
             let x = x - r;
             #approx * mul
         }
@@ -30,7 +34,7 @@ pub fn gen_exp2(num_terms: usize, num_bits: usize) -> TokenStream {
 }
 
 pub fn gen_exp(_num_terms: usize, num_bits: usize) -> TokenStream {
-    let fty = format_ident!("f{}", num_bits);
+    let fty = helpers::get_fty(num_bits);
     quote!(
         fn exp(x: #fty) -> #fty {
             exp2(x * std::#fty::consts::LOG2_E)
@@ -39,8 +43,11 @@ pub fn gen_exp(_num_terms: usize, num_bits: usize) -> TokenStream {
 }
 
 pub fn gen_exp_m1(num_terms: usize, num_bits: usize) -> TokenStream {
-    let fty = format_ident!("f{}", num_bits);
-    let suffix = format!("f{}", num_bits);
+    let fty = helpers::get_fty(num_bits);
+    let uty = helpers::get_uty(num_bits);
+    let suffix = helpers::get_suffix(num_bits);
+    let one = helpers::get_one(num_bits);
+    let escale = helpers::get_escale(num_bits);
 
     let xmin = -0.5;
     let xmax = 0.5;
@@ -56,7 +63,7 @@ pub fn gen_exp_m1(num_terms: usize, num_bits: usize) -> TokenStream {
         fn exp_m1(x: #fty) -> #fty {
             let x = x * std::#fty::consts::LOG2_E;
             let r = x.round();
-            let mul = #fty::from_bits((r.mul_add(0x00800000 as #fty, 0x3f800000 as #fty)) as u32);
+            let mul = #fty::from_bits((r.mul_add(#escale as #fty, #one as #fty)) as #uty);
             let x = x - r;
             #approx * mul + (mul - 1.0)
         }
@@ -64,8 +71,8 @@ pub fn gen_exp_m1(num_terms: usize, num_bits: usize) -> TokenStream {
 }
 
 pub fn gen_ln_1p(num_terms: usize, num_bits: usize) -> TokenStream {
-    let fty = format_ident!("f{}", num_bits);
-    let suffix = format!("f{}", num_bits);
+    let fty = helpers::get_fty(num_bits);
+    let suffix = helpers::get_suffix(num_bits);
 
     let xmin = 0.0;
     let xmax = 1.0;
@@ -88,8 +95,10 @@ pub fn gen_ln_1p(num_terms: usize, num_bits: usize) -> TokenStream {
 }
 
 pub fn gen_log2(num_terms: usize, num_bits: usize) -> TokenStream {
-    let fty = format_ident!("f{}", num_bits);
-    let suffix = format!("f{}", num_bits);
+    let fty = helpers::get_fty(num_bits);
+    let suffix = helpers::get_suffix(num_bits);
+    let one = helpers::get_one(num_bits);
+    let escale = helpers::get_escale(num_bits);
 
     let xmin = -0.5;
     let xmax = 0.5;
@@ -104,7 +113,7 @@ pub fn gen_log2(num_terms: usize, num_bits: usize) -> TokenStream {
     quote!(
         fn log2(x: #fty) -> #fty {
             let exponent = (x.to_bits() >> 23) as i32 - 0x7f;
-            let x = #fty::from_bits((x.to_bits() & 0x7fffff) | 0x3f800000) - 1.5;
+            let x = #fty::from_bits((x.to_bits() & (#escale-1)) | #one) - 1.5;
             let y: #fty = #approx;
             y + (exponent as #fty)
         }
@@ -112,7 +121,7 @@ pub fn gen_log2(num_terms: usize, num_bits: usize) -> TokenStream {
 }
 
 pub fn gen_ln(_num_terms: usize, num_bits: usize) -> TokenStream {
-    let fty = format_ident!("f{}", num_bits);
+    let fty = helpers::get_fty(num_bits);
 
     quote!(
         fn ln(x: #fty) -> #fty {
@@ -122,7 +131,7 @@ pub fn gen_ln(_num_terms: usize, num_bits: usize) -> TokenStream {
 }
 
 pub fn gen_log10(_num_terms: usize, num_bits: usize) -> TokenStream {
-    let fty = format_ident!("f{}", num_bits);
+    let fty = helpers::get_fty(num_bits);
 
     quote!(
         fn log10(x: #fty) -> #fty {
@@ -132,7 +141,7 @@ pub fn gen_log10(_num_terms: usize, num_bits: usize) -> TokenStream {
 }
 
 pub fn gen_log(_num_terms: usize, num_bits: usize) -> TokenStream {
-    let fty = format_ident!("f{}", num_bits);
+    let fty = helpers::get_fty(num_bits);
 
     quote!(
         fn log(x: #fty, base: #fty) -> #fty {
@@ -142,7 +151,7 @@ pub fn gen_log(_num_terms: usize, num_bits: usize) -> TokenStream {
 }
 
 pub fn gen_powf(_num_terms: usize, num_bits: usize) -> TokenStream {
-    let fty = format_ident!("f{}", num_bits);
+    let fty = helpers::get_fty(num_bits);
 
     quote!(
         fn powf(x: #fty, y: #fty) -> #fty {
@@ -152,7 +161,7 @@ pub fn gen_powf(_num_terms: usize, num_bits: usize) -> TokenStream {
 }
 
 pub fn gen_powi(_num_terms: usize, num_bits: usize) -> TokenStream {
-    let fty = format_ident!("f{}", num_bits);
+    let fty = helpers::get_fty(num_bits);
 
     // Note, for constant values under 16, the code path is very short.
     quote!(
@@ -195,11 +204,12 @@ pub fn gen_log_exp(num_bits: usize) -> (TokenStream, TokenStream) {
     let powi = gen_powi(16, num_bits);
 
     let bit = (2.0_f64).powi(if num_bits == 32 { 23 } else { 52 });
+    let fty = helpers::get_fty(num_bits);
 
     let test_exp_a = gen_test(
         quote!(test_exp_a),
         quote!(x.exp()),
-        quote!(exp(x as f32) as f64),
+        quote!(exp(x as #fty) as f64),
         bit * 3.0,
         0.0,
         1.0,
@@ -207,7 +217,7 @@ pub fn gen_log_exp(num_bits: usize) -> (TokenStream, TokenStream) {
     let test_exp_b = gen_test(
         quote!(test_exp_b),
         quote!(x.exp()),
-        quote!(exp(x as f32) as f64),
+        quote!(exp(x as #fty) as f64),
         bit * 10.0,
         1.0,
         2.0,
@@ -215,7 +225,7 @@ pub fn gen_log_exp(num_bits: usize) -> (TokenStream, TokenStream) {
     let test_exp_m1 = gen_test(
         quote!(test_exp_m1),
         quote!(x.exp_m1()),
-        quote!(exp_m1(x as f32) as f64),
+        quote!(exp_m1(x as #fty) as f64),
         bit * 3.0,
         0.0,
         1.0,
@@ -223,7 +233,7 @@ pub fn gen_log_exp(num_bits: usize) -> (TokenStream, TokenStream) {
     let test_exp2 = gen_test(
         quote!(test_exp2),
         quote!(x.exp2()),
-        quote!(exp2(x as f32) as f64),
+        quote!(exp2(x as #fty) as f64),
         bit * 2.0,
         0.0,
         1.0,
@@ -231,7 +241,7 @@ pub fn gen_log_exp(num_bits: usize) -> (TokenStream, TokenStream) {
     // let test_exp2_x = gen_test(
     //     quote!(test_exp2_x),
     //     quote!(x.exp2()),
-    //     quote!(exp2_approx(x as f32) as f64),
+    //     quote!(exp2_approx(x as #fty) as f64),
     //     0.05,
     //     0.0,
     //     1.0,
@@ -240,7 +250,7 @@ pub fn gen_log_exp(num_bits: usize) -> (TokenStream, TokenStream) {
     let test_ln = gen_test(
         quote!(test_ln),
         quote!(x.ln()),
-        quote!(ln(x as f32) as f64),
+        quote!(ln(x as #fty) as f64),
         bit * 2.0,
         1.0,
         std::f64::consts::E,
@@ -248,7 +258,7 @@ pub fn gen_log_exp(num_bits: usize) -> (TokenStream, TokenStream) {
     let test_ln_1p_a = gen_test(
         quote!(test_ln_1p_a),
         quote!(x.ln_1p()),
-        quote!(ln_1p(x as f32) as f64),
+        quote!(ln_1p(x as #fty) as f64),
         bit * 2.0,
         0.0,
         1.0,
@@ -256,7 +266,7 @@ pub fn gen_log_exp(num_bits: usize) -> (TokenStream, TokenStream) {
     let test_ln_1p_b = gen_test(
         quote!(test_ln_1p_b),
         quote!(x.ln_1p()),
-        quote!(ln_1p(x as f32) as f64),
+        quote!(ln_1p(x as #fty) as f64),
         bit * 3.0,
         1.0,
         std::f64::consts::E * 3.0 - 1.0,
@@ -264,7 +274,7 @@ pub fn gen_log_exp(num_bits: usize) -> (TokenStream, TokenStream) {
     let test_log2 = gen_test(
         quote!(test_log2),
         quote!(x.log2()),
-        quote!(log2(x as f32) as f64),
+        quote!(log2(x as #fty) as f64),
         bit * 2.0,
         0.25,
         4.25,
@@ -272,7 +282,7 @@ pub fn gen_log_exp(num_bits: usize) -> (TokenStream, TokenStream) {
     let test_log10 = gen_test(
         quote!(test_log10),
         quote!(x.log10()),
-        quote!(log10(x as f32) as f64),
+        quote!(log10(x as #fty) as f64),
         bit * 2.0,
         0.1,
         10.1,
@@ -280,7 +290,7 @@ pub fn gen_log_exp(num_bits: usize) -> (TokenStream, TokenStream) {
     let test_log_2 = gen_test(
         quote!(test_log_2),
         quote!(x.log(2.0)),
-        quote!(log(x as f32, 2.0) as f64),
+        quote!(log(x as #fty, 2.0) as f64),
         bit * 2.0,
         0.5,
         1.5,
@@ -288,7 +298,7 @@ pub fn gen_log_exp(num_bits: usize) -> (TokenStream, TokenStream) {
     let test_log_e = gen_test(
         quote!(test_log_e),
         quote!(x.log(std::f64::consts::E)),
-        quote!(log(x as f32, std::f32::consts::E) as f64),
+        quote!(log(x as #fty, std::f64::consts::E as #fty) as f64),
         bit * 2.0,
         0.5,
         1.5,
@@ -297,7 +307,7 @@ pub fn gen_log_exp(num_bits: usize) -> (TokenStream, TokenStream) {
     let test_powf_2 = gen_test(
         quote!(test_powf_2),
         quote!(x.powf(2.0)),
-        quote!(powf(x as f32, 2.0) as f64),
+        quote!(powf(x as #fty, 2.0) as f64),
         bit * 4.0,
         0.5,
         1.5,
@@ -305,7 +315,7 @@ pub fn gen_log_exp(num_bits: usize) -> (TokenStream, TokenStream) {
     let test_powf_m1 = gen_test(
         quote!(test_powf_m1),
         quote!(x.powf(-1.0)),
-        quote!(powf(x as f32, -1.0) as f64),
+        quote!(powf(x as #fty, -1.0) as f64),
         bit * 4.0,
         0.5,
         1.5,
@@ -314,7 +324,7 @@ pub fn gen_log_exp(num_bits: usize) -> (TokenStream, TokenStream) {
     let test_powi_2 = gen_test(
         quote!(test_powi_2),
         quote!(x.powi(2)),
-        quote!(powi(x as f32, 2) as f64),
+        quote!(powi(x as #fty, 2) as f64),
         bit * 2.0,
         0.5,
         1.5,
@@ -322,7 +332,7 @@ pub fn gen_log_exp(num_bits: usize) -> (TokenStream, TokenStream) {
     let test_powi_3 = gen_test(
         quote!(test_powi_3),
         quote!(x.powi(3)),
-        quote!(powi(x as f32, 3) as f64),
+        quote!(powi(x as #fty, 3) as f64),
         bit * 4.0,
         0.12,
         1.2,
@@ -330,7 +340,7 @@ pub fn gen_log_exp(num_bits: usize) -> (TokenStream, TokenStream) {
     let test_powi_m1 = gen_test(
         quote!(test_powi_m1),
         quote!(x.powi(-1)),
-        quote!(powi(x as f32, -1) as f64),
+        quote!(powi(x as #fty, -1) as f64),
         bit * 2.0,
         0.5,
         1.5,
@@ -338,7 +348,7 @@ pub fn gen_log_exp(num_bits: usize) -> (TokenStream, TokenStream) {
     let test_powi_m2 = gen_test(
         quote!(test_powi_m2),
         quote!(x.powi(-2)),
-        quote!(powi(x as f32, -2) as f64),
+        quote!(powi(x as #fty, -2) as f64),
         bit * 6.0,
         0.5,
         1.5,
@@ -346,7 +356,7 @@ pub fn gen_log_exp(num_bits: usize) -> (TokenStream, TokenStream) {
     let test_powi_16 = gen_test(
         quote!(test_powi_16),
         quote!(x.powi(16)),
-        quote!(powi(x as f32, 16) as f64),
+        quote!(powi(x as #fty, 16) as f64),
         bit * 7.0,
         0.25,
         1.0,
