+++ src/bin/libmgen/main.rs
@@ -8,6 +8,7 @@ mod log_exp;
 mod recip_sqrt;
 mod test;
 mod trig;
+mod helpers;
 
 use hyperbolic::*;
 use inv_trig::*;
@@ -19,7 +20,6 @@ use auxfuncs::*;
 fn generate_libm(path: &str, num_bits: usize) -> std::io::Result<()> {
     let mut file = std::fs::File::create(path)?;
 
-    // write!(file, "\n{}\n", gen_sin(16, false))?;
     let (trig, trig_tests) = gen_quadrant_trig(num_bits);
     let (inv_trig, inv_trig_tests) = gen_inv_trig(num_bits);
     let (log_exp, log_exp_tests) = gen_log_exp(num_bits);
@@ -48,11 +48,13 @@ fn generate_libm(path: &str, num_bits: usize) -> std::io::Result<()> {
     file.write_all(functions.to_string().as_bytes())?;
     file.write_all(tests.to_string().as_bytes())?;
 
-    //write!(file, "\n{}\n", gen_test("powi_16", quote!(x.powi(16)), quote!(powf(x as f32, 16.0) as f64), bit*6.0, 0.5, 1.5))?;
-
     Ok(())
 }
 
+
 fn main() {
-    generate_libm("tests/libm.rs", 32).unwrap();
+    let val = doctor_syn::expr!((2.16065388452409390396).cos()).eval(60).unwrap();
+    println!("val={}", val);
+    generate_libm("tests/libm32.rs", 32).unwrap();
+    generate_libm("tests/libm64.rs", 64).unwrap();
 }
